requirements.txt
# Web Framework
Flask==2.3.3
gunicorn==21.2.0

# Firebase Admin SDK
firebase-admin==6.2.0

# Slack SDK
slack-sdk==3.21.3

# Scraping
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# YAML Processing
PyYAML==6.0.1

# Async Support (現在は同期化したが依存は残しておく)
aiohttp==3.8.5
asyncio==3.4.3

# Date/Time Handling
python-dateutil==2.8.2
pytz==2023.3

# Cloud Services
google-cloud-secret-manager==2.16.3
google-cloud-logging==3.5.0

# Testing
pytest==7.4.2
pytest-asyncio==0.21.1
pytest-mock==3.11.1

# Development Tools
black==23.7.0
flake8==6.1.0
mypy==1.5.1
isort==5.12.0

# Monitoring & Logging
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-instrumentation-flask==0.41b0

# Type Hints
types-PyYAML==6.0.12.11
types-requests==2.31.0.2

# Security
cryptography==41.0.3

# Performance
cachetools==5.3.1
